Only load completed candles
This function downloads data and saves it to a file which is then loaded into
the database using mariadb's LOAD DATA INFILE.

Crucially, the data is loaded into the database using LOAD DATA INFILE IGNORE
meaning if the candle already exist, it is ignored.  What was happening before
this change was data would be inserted early when the candle was incomplete,
and later when the candle completed and ie, had different high/low/close/volume
values, the latest correct values would be ignored and not inserted into the
database.

That problem could have been solved using different approaches, ie:

  * Use LOAD DATA INFILE REPLACE
    I decided against this because it represents additional write operations
    which lead to dirty memory buffers and generally poorer performance.

  * Only insert complete candles
    This is the approach implemented here.  The downside to this is that data
    will be slightly delayed, as we wait for the candle to complete before
    storing it.  The benefit is that data will be correct.  The downside can
    be minimized by using shorter timeframes.  At the moment I'm using minute
    timeframe, but this instruments API supports down to 5second granularity.