Use whatever timeframe order was passed in the command line
Sorting the timeframes would make sense if this was being used to
calculate synthetics, because the larger timeframes would be calculated
based on the smallest timeframes, hence it made sense to update the
smallest timeframes first.

However, this value is not actually used to calculate synthetic timeframes
so I can see no reason to sort it, and it's misleading to sort it starts
downloading data in a different order than what I asked for in the command
line, hence I decided to remove the sort.